+++ examples/map_milky_way_stars.py
@@ -1,3 +1,5 @@
+from PIL import Image, ImageFilter
+
 from starplot import MapPlot, Mollweide, _
 from starplot.data.catalogs import BIG_SKY
 from starplot.styles import PlotStyle, extensions
@@ -10,38 +12,23 @@ style = PlotStyle().extend(
 
 _sizer = size_by_magnitude_factory(6, 0.02, 7)
 
-
-def alpha(s):
-    if s.magnitude > 9:
-        return 0.5
-    else:
-        return 0.9
-
-
 p = MapPlot(
     projection=Mollweide(),
     style=style,
     resolution=4800,
 )
-
 p.stars(
     where=[_.magnitude < 11],
     where_labels=[False],
     size_fn=_sizer,
-    alpha_fn=alpha,
+    alpha_fn=lambda s: 0.95 if s.magnitude < 9 else 0.6,
     color_fn=color_by_bv,
     catalog=BIG_SKY,
-    style__marker__edge_color="#c5c5c5",
+    style__marker__edge_color="#fff",
 )
-p.stars(
-    where=[_.magnitude < 6],
-    where_labels=[False],
-    size_fn=lambda s: _sizer(s) * 1.5,
-    alpha_fn=lambda s: 0.4,
-    color_fn=color_by_bv,
-    catalog=BIG_SKY,
-    style__marker__symbol="star_8",
-    style__marker__edge_color=None,
-)
-
 p.export("map_milky_way_stars.png", padding=0.1, transparent=True)
+
+# apply a median filter and increase contrast
+with Image.open("map_milky_way_stars.png") as img:
+    filtered = img.filter(ImageFilter.MedianFilter(size=5))
+    filtered.save("map_milky_way_stars.png")