+++ src/starplot/plots/base.py
@@ -898,11 +898,13 @@ class BasePlot(DebugPlotterMixin, TextPlotterMixin, ABC):
                 style=style,
             )
 
+    @profile
     @use_style(PathStyle, "ecliptic")
     def ecliptic(
         self,
         style: PathStyle = None,
         label: str = "ECLIPTIC",
+        num_labels: int = 2,
         collision_handler: CollisionHandler = None,
     ):
         """Plots the ecliptic
@@ -910,6 +912,7 @@ class BasePlot(DebugPlotterMixin, TextPlotterMixin, ABC):
         Args:
             style: Styling of the ecliptic. If None, then the plot's style will be used
             label: How the ecliptic will be labeled on the plot
+            num_labels: Max number of labels to plot along the line
             collision_handler: An instance of [CollisionHandler][starplot.CollisionHandler] that describes what to do on label collisions with other labels, markers, etc. If `None`, then the collision handler of the plot will be used.
         """
         x = []
@@ -925,34 +928,23 @@ class BasePlot(DebugPlotterMixin, TextPlotterMixin, ABC):
             if self.in_bounds(ra * 15, dec):
                 inbounds.append((ra * 15, dec))
 
-        self.ax.plot(
-            x,
-            y,
-            dash_capstyle=style.line.dash_capstyle,
-            clip_path=self._background_clip_path,
-            gid="ecliptic-line",
-            **style.line.matplot_kwargs(self.scale),
-            **self._plot_kwargs(),
+        coords = [(ra * 15, dec) for ra, dec in ecliptic.RA_DECS]
+
+        self.line_label(
+            style=style,
+            label=label.upper(),
+            num_labels=num_labels,
+            collision_handler=collision_handler or self.collision_handler,
+            coordinates=coords,
         )
 
-        if label and len(inbounds) > 4:
-            label_spacing = int(len(inbounds) / 4)
-
-            for ra, dec in [inbounds[label_spacing], inbounds[label_spacing * 2]]:
-                self.text(
-                    label,
-                    ra,
-                    dec,
-                    style.label,
-                    collision_handler=collision_handler or self.collision_handler,
-                    gid="ecliptic-label",
-                )
-
+    @profile
     @use_style(PathStyle, "celestial_equator")
     def celestial_equator(
         self,
         style: PathStyle = None,
         label: str = "CELESTIAL EQUATOR",
+        num_labels: int = 2,
         collision_handler: CollisionHandler = None,
     ):
         """
@@ -961,37 +953,83 @@ class BasePlot(DebugPlotterMixin, TextPlotterMixin, ABC):
         Args:
             style: Styling of the celestial equator. If None, then the plot's style will be used
             label: How the celestial equator will be labeled on the plot
+            num_labels: Max number of labels to plot along the line
             collision_handler: An instance of [CollisionHandler][starplot.CollisionHandler] that describes what to do on label collisions with other labels, markers, etc. If `None`, then the collision handler of the plot will be used.
         """
-        x = []
-        y = []
-
-        # TODO : handle wrapping
-
         label = translate(label, self.language)
+        coords = [(ra, 0) for ra in range(0, 361)]
+        self.line_label(
+            style=style,
+            label=label.upper(),
+            num_labels=num_labels,
+            collision_handler=collision_handler or self.collision_handler,
+            coordinates=coords,
+            gid="celestial-equator",
+        )
 
-        for ra in range(25):
-            x0, y0 = self._prepare_coords(ra * 15, 0)
-            x.append(x0)
-            y.append(y0)
+    @use_style(PathStyle)
+    def line_label(
+        self,
+        style: PathStyle,
+        label: str = None,
+        collision_handler: CollisionHandler = None,
+        num_labels: int = 2,
+        coordinates: list[tuple[float, float]] = None,
+        geometry: LineString = None,
+        **kwargs,
+    ):
+        """Plots a line, with optional labels. Either coordinates OR geometry must be specified.
+
+        Args:
+            style: Style of the line
+            label: Label for the line
+            num_labels: Number of labels to plot along the line
+            coordinates: List of coordinates, e.g. `[(ra, dec), (ra, dec)]`
+            geometry: A shapely LineString. If this value is passed, then the `coordinates` kwarg will be ignored.
+
+        """
+
+        if coordinates is None and geometry is None:
+            raise ValueError("Must pass coordinates or geometry when plotting lines.")
+
+        coords = geometry.coords if geometry is not None else coordinates
+        prepared_coords = [self._prepare_coords(*p) for p in coords]
+        x, y = zip(*prepared_coords)
+
+        gid = kwargs.get("gid") or "line"
 
         self.ax.plot(
             x,
             y,
+            clip_on=True,
             clip_path=self._background_clip_path,
-            gid="celestial-equator-line",
+            dash_capstyle=style.line.dash_capstyle,
+            gid=gid,
             **style.line.matplot_kwargs(self.scale),
             **self._plot_kwargs(),
         )
 
-        if label:
-            label_spacing = (self.ra_max - self.ra_min) / 3
-            for ra in np.arange(self.ra_min, self.ra_max, label_spacing):
-                self.text(
-                    label,
-                    ra,
-                    0.25,
-                    style.label,
-                    collision_handler=collision_handler or self.collision_handler,
-                    gid="celestial-equator-label",
-                )
+        if not label:
+            return
+
+        prepared_coords = [
+            (x, y) for x, y in prepared_coords if self._in_bounds_xy(x, y)
+        ]
+
+        if not prepared_coords:
+            return
+
+        x, y = zip(*prepared_coords)
+
+        self._text_line(
+            x,
+            y,
+            label,
+            num_labels=num_labels,
+            collision_handler=collision_handler or self.collision_handler,
+            min_spacing=0.65,
+            **style.label.matplot_kwargs(self.scale),
+            **self._plot_kwargs(),
+            clip_path=self._background_clip_path,
+            gid=gid,
+        )